requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-jet==1.0.8
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mysqlclient==2.0.3
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
